Disable custom baked block models for modules if Sodium is present. It completely replaces the render pipeline but does not re-implement FabricBakedModel facilities, so these cannot work. Means we also have to disable the facade module. It gets hidden in the creative tab and in REI via a plugin. See #171.

+++ fabric/src/main/java/li/cil/tis3d/client/fabric/ClientBootstrapFabric.java
@@ -12,13 +12,18 @@ import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientChunkEvents;
 import net.fabricmc.fabric.api.client.model.ModelLoadingRegistry;
 import net.fabricmc.fabric.api.event.client.ClientSpriteRegistryCallback;
 import net.fabricmc.fabric.api.event.client.player.ClientPickBlockGatherCallback;
+import net.fabricmc.loader.api.FabricLoader;
 import net.fabricmc.loader.impl.entrypoint.EntrypointUtils;
 import net.minecraft.world.inventory.InventoryMenu;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.phys.BlockHitResult;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 public final class ClientBootstrapFabric implements ClientModInitializer {
+    private static final Logger LOGGER = LogManager.getLogger();
+
     @Override
     public void onInitializeClient() {
         ClientBootstrap.run();
@@ -51,6 +56,10 @@ public final class ClientBootstrapFabric implements ClientModInitializer {
         ClientSpriteRegistryCallback.event(InventoryMenu.BLOCK_ATLAS).register((atlasTexture, registry) ->
             Textures.visitBlockAtlasTextures(registry::register));
 
-        ModelLoadingRegistry.INSTANCE.registerResourceProvider(rm -> new ModuleModelLoader());
+        if (FabricLoader.getInstance().isModLoaded("sodium")) {
+            LOGGER.warn("Sodium detected, disabling modules that need custom block model rendering. See https://github.com/MightyPirates/TIS-3D/issues/171");
+        } else {
+            ModelLoadingRegistry.INSTANCE.registerResourceProvider(rm -> new ModuleModelLoader());
+        }
     }
 }
